Min css and js are now used
adding a new line to min.js fixes bug #2

css is now uglified as well

+++ src/build.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 # this script will compile the vis library into a tw-plugin
 
-distPath="../dist/felixhayashi/vis/"    # Path to dist-files
 modulePath="$:/plugins/felixhayashi/vis"  # module path
-cssSrc="vis/vis.css"                    # path to the css
-cssOut="$distPath/files/vis.css"        # outpath of the css
-cd vis; images=(img/*/*); cd ..         # array of images used by vis relative to css dir
+cssSrc="vis/vis.css"                      # path to the css within the source dir
+distPath="../dist/felixhayashi/vis/"      # Path to dist-files
+cssOut="$distPath/files/vis.css"          # outpath of the css in the dist dir
+cd vis; images=(img/*/*); cd ..           # array of images used by vis relative to css dir
 
-createTWFilesFile() {
+buildTWFilesFile() {
   
 # print tiddlywiki.files preamble
 printf '{
@@ -41,17 +41,11 @@ printf '
         "module-type": "library"
       }}
   ]
-}' "$modulePath" "$modulePath" >> "$distPath/files/tiddlywiki.files"
+}' "$modulePath" "$modulePath" >> "$distPath/files/tiddlywiki.files" # can't access arguments by index in gnu's printf
 
 }
 
-integrateCSS() {
-
-  # insert macro at top
-  echo \
-'\define datauri(title)
-<$macrocall $name="makedatauri" type={{$title$!!type}} text={{$title$}}/>
-\end'$'\n' > $cssOut
+buildStyles() {
   
   # replace urls
   gawk -v mpath="$modulePath" '
@@ -59,20 +53,42 @@ integrateCSS() {
       pos = match($0, /(.*)[\"'\''](.*)[\"'\''](.*)/, arr);
       if(pos != 0) print arr[1] "<<datauri \"" mpath "/" arr[2] "\" >>" arr[3]
       else print
-    }' $cssSrc >> $cssOut
+    }' $cssSrc > $cssOut
+    
+  # uglify; redirect stdin so its not closed by npm command
+  body=$(uglifycss $cssOut < /dev/null)
+  
+  # insert the macro function at the top
+  echo \
+'\define datauri(title)
+<$macrocall $name="makedatauri" type={{$title$!!type}} text={{$title$}}/>
+\end'$'\n'$'\n'$body > $cssOut
     
 }
 
+copyMedia() {
+  # copy images
+  cp -r ./vis/img/ "$distPath/files/img"
+}
+
+buildScripts() {
+  cp ./vis/vis.min.js "$distPath/files/vis.js"
+  # append empty line
+  # fix for https://github.com/felixhayashi/tw-vis/issues/2
+  echo "" >> "$distPath/files/vis.js"
+}
+
 createStructure() {
-  mkdir -p "$distPath"
+  [ -d "$distPath" ] && rm -rf "$distPath"
+  mkdir -p "$distPath/files"
   cp ./plugin.info "$distPath"
-  mkdir "$distPath/files/"
-  cp -r ./vis/* "$distPath/files/"
 }
 
 #execute
 
-rm -rf "$distPath"
 createStructure
-integrateCSS
-createTWFilesFile
+copyMedia
+buildStyles
+buildScripts
+buildTWFilesFile
+
